Add BinaryExpressionExtension
